Distribute the initial states better
Vary the starting value, modulo the total number of values. This
eliminates a bias where the values are just increasing which,
unfortunately, as we aggregate more and more states creates too much
variation. This technique results in a flatter profile on aggregate and
is a better representation for partition maps.